services/user_service/requirements.txt
grpcio
grpcio-tools
protobuf
SQLAlchemy
psycopg2-binary